refactor: 문자열 계산기 소스 코드 구조 변경 2
