--- CalculatorEx/src/main/java/org/example/InputException.java
@@ -1,19 +1,0 @@
-package org.example;
-
-public class InputException extends RuntimeException {
-    public InputException() {
-        super();
-    }
-
-    public InputException(String message) {
-        super(message);
-    }
-
-    public InputException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public InputException(Throwable cause) {
-        super(cause);
-    }
-}
