+++ CalculatorEx/src/test/java/org/example/StringToIntTest.java
@@ -1,0 +1,19 @@
+package org.example;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+public class StringToIntTest {
+    @Test
+    public void stringToIntTest() throws Exception {
+        // given
+        String s = ",";
+
+        // when
+
+        // then
+        assertThrows(RuntimeException.class, () -> Integer.parseInt(s));
+    }
+}
